Store `heightCache` Map inside the component state
A previous *hack* was more efficient.
Not sure if this is the best possible approach,
but it is probably the choicest compromise
between performance and code clarity.

We used to increment a dummy state value
to trigger the list rerender.
Currently, we are using Map cloning, which is O(n).